REORGANIZE: Added static static functions
